Reorganized attalos/imgtxt_algorithms/approaches/ folder.
