+++ attalos/imgtxt_algorithms/approaches/base.py
