+++ attalos/imgtxt_algorithms/approaches/fast0tag.py
