+++ attalos/imgtxt_algorithms/approaches/model.py
