+++ attalos/imgtxt_algorithms/approaches/multihot.py
@@ -1,7 +1,7 @@
 import numpy as np
 import tensorflow as tf
 
-from attalos.imgtxt_algorithms.approaches.attalos_model import AttalosModel
+from attalos.imgtxt_algorithms.approaches.base import AttalosModel
 from attalos.util.transformers import OneHot
 
 
@@ -9,6 +9,7 @@ class MultihotModel(AttalosModel):
     """
     This model performs logistic regression via NN using multihot vectors as targets.
     """
+
     def _construct_model_info(self, input_size, output_size, learning_rate):
         model_info = {}
         model_info["input"] = tf.placeholder(shape=(None, input_size), dtype=tf.float32)
@@ -20,17 +21,68 @@ class MultihotModel(AttalosModel):
         model_info["optimizer"] = tf.train.AdamOptimizer(learning_rate=learning_rate).minimize(model_info["loss"])
         return model_info
     
-    def __init__(self, wv_model, train_dataset, test_dataset, **kwargs):
-        self.cross_eval = kwargs.get("cross_eval", False)
-        self.one_hot = OneHot([train_dataset] if self.cross_eval else [train_dataset, test_dataset], 
-                              valid_vocab=wv_model.vocab)
+    def __init__(self, wv_model, datasets, **kwargs): #train_dataset, test_dataset, **kwargs):
+        #self.cross_eval = kwargs.get("cross_eval", False)
+        self.wv_model = wv_model
+        #self.one_hot = OneHot([train_dataset] if self.cross_eval else [train_dataset, test_dataset],
+        #                      valid_vocab=wv_model.vocab)
+        self.one_hot = OneHot(datasets, valid_vocab=wv_model.vocab)
+        train_dataset = datasets[0]  # train_dataset should always be first in datasets iterable
         self.learning_rate = kwargs.get("learning_rate", 0.0001)
         self.model_info = self._construct_model_info(
                 input_size = train_dataset.img_feat_size,
                 output_size = self.one_hot.vocab_size, 
                 learning_rate = self.learning_rate
         )
+        self.test_one_hot = None
         super(MultihotModel, self).__init__()
+
+    def prep_fit(self, data):
+        img_feats_list, text_feats_list = data
+
+        img_feats = np.array(img_feats_list)
+        text_feats = [self.one_hot.get_multiple(text_feats) for text_feats in text_feats_list]
+        text_feats = np.array(text_feats)
+
+        fetches = [self.model_info["optimizer"], self.model_info["loss"]]
+        feed_dict = {
+            self.model_info["input"]: img_feats,
+            self.model_info["y_truth"]: text_feats
+        }
+        return fetches, feed_dict
+
+    def prep_predict(self, dataset, cross_eval=False):
+        if cross_eval:
+            raise Exception("cross_eval is True. Multihot model does not support cross evaluation.")
+        x = []
+        y = []
+        for idx in dataset:
+            image_feats, text_feats = dataset.get_index(idx)
+            text_feats = self.one_hot.get_multiple(text_feats)
+            x.append(image_feats)
+            y.append(text_feats)
+        x = np.asarray(x)
+        y = np.asarray(y)
+
+        fetches = [self.model_info["predictions"],]
+        feed_dict = {
+            self.model_info["input"]: x
+        }
+        truth = y
+        return fetches, feed_dict, truth
+
+    def post_predict(self, predict_fetches, cross_eval=False):
+        if cross_eval:
+            raise Exception("cross_eval is True. Multihot model does not support cross evaluation.")
+        predictions = predict_fetches[0]
+        return predictions
+
+    def get_training_loss(self, fit_fetches):
+        _, loss = fit_fetches
+        return loss
+
+
+    """
 
     # is a generator
     def iter_batches(self, dataset, batch_size):
@@ -82,4 +134,5 @@ class MultihotModel(AttalosModel):
             predictions = fetches
         return (predictions, loss) if y else predictions
 
+    """
